Reorganizes Lovell files/structure
