+++ src/lib/drupal/staticProps.ts
@@ -4,30 +4,23 @@ import { QueryOpts } from 'next-drupal-query'
 import { drupalClient } from '@/lib/drupal/drupalClient'
 import { queries } from '@/data/queries'
 import {
-  LISTING_RESOURCE_TYPE_URL_SEGMENTS,
   ListingPageStaticPropsContextProps,
-  ListingResourceTypeType,
   getListingPageStaticPropsContext,
-} from './listingPages'
+} from '@/lib/drupal/listingPages'
 import {
-  LOVELL,
-  LovellStaticPropsContextProps,
+  getLovellStaticPropsResource,
   getLovellStaticPropsContext,
-  LovellFormattedResource,
-  LovellStaticPropsResource,
+} from '@/lib/drupal/lovell/staticProps'
+import {
   LovellResourceType,
-  isLovellBifurcatedResource,
-  getLovellVariantOfUrl,
-  getLovellChildVariantOfResource,
-  LovellBifurcatedFormattedResource,
-  LovellListingPageFormattedResource,
-  getOppositeChildVariant,
-} from './lovell'
+  LovellStaticPropsContextProps,
+  LovellStaticPropsResource,
+  LovellFormattedResource,
+} from './lovell/types'
 import { FormattedResource } from '@/data/queries'
 import { RESOURCE_TYPES, ResourceTypeType } from '@/lib/constants/resourceTypes'
 import { ListingPageDataOpts } from '@/lib/drupal/listingPages'
 import { NewsStoryDataOpts } from '@/data/queries/newsStory'
-import { PAGE_SIZES } from '../constants/pageSizes'
 
 export type ExpandedStaticPropsContext = GetStaticPropsContext & {
   path: string
@@ -130,98 +123,6 @@ export async function getDefaultStaticPropsResource(
   const id = pathInfo.entity?.uuid
   const queryOpts = getStaticPropsQueryOpts(resourceType, id, context)
   return fetchSingleStaticPropsResource(resourceType, pathInfo, queryOpts)
-}
-
-export async function getLovellListingPageStaticPropsResource(
-  resourceType: ListingResourceTypeType,
-  pathInfo: DrupalTranslatedPath,
-  context: ExpandedStaticPropsContext
-): Promise<LovellStaticPropsResource<LovellListingPageFormattedResource>> {
-  const id = pathInfo.entity?.uuid
-  const childVariantPage = (await fetchSingleStaticPropsResource(
-    resourceType,
-    pathInfo,
-    {
-      id,
-      // Do not pass a page number; we need all of the pages
-      // so we can merge and then calculate page data
-    }
-  )) as LovellListingPageFormattedResource
-
-  const federalPagePathInfo = await drupalClient.translatePath(
-    getLovellVariantOfUrl(context.drupalPath, LOVELL.federal.variant)
-  )
-  if (!federalPagePathInfo) {
-    return childVariantPage
-  }
-  const federalPageId = federalPagePathInfo.entity?.uuid
-  const federalPage = (await fetchSingleStaticPropsResource(
-    resourceType,
-    federalPagePathInfo,
-    {
-      id: federalPageId,
-      // Again, do not pass specific page number
-    }
-  )) as LovellListingPageFormattedResource
-
-  const itemProp = LISTING_RESOURCE_TYPE_URL_SEGMENTS[resourceType]
-  const allMergedItems = [
-    ...childVariantPage[itemProp],
-    ...federalPage[itemProp],
-  ]
-  const pageSize = PAGE_SIZES[resourceType]
-  const sliceStart = (context.listing.page - 1) * pageSize
-  const sliceEnd = sliceStart + pageSize
-  const pagedMergedItems = allMergedItems.slice(sliceStart, sliceEnd)
-  const totalItems = childVariantPage.totalItems + federalPage.totalItems
-  const totalPages = Math.ceil(totalItems / pageSize) || 0
-
-  return {
-    ...childVariantPage,
-    [itemProp]: pagedMergedItems,
-    currentPage: context.listing.page,
-    totalItems,
-    totalPages,
-    canonicalLink: childVariantPage.entityPath,
-    lovellVariant: context.lovell.variant,
-    lovellSwitchPath: getLovellVariantOfUrl(
-      childVariantPage.entityPath,
-      getOppositeChildVariant(context.lovell.variant)
-    ),
-  }
-}
-
-export async function getLovellStaticPropsResource(
-  resourceType: LovellResourceType,
-  pathInfo: DrupalTranslatedPath,
-  context: ExpandedStaticPropsContext
-): Promise<LovellStaticPropsResource<LovellFormattedResource>> {
-  // Lovell listing pages need Federal items merged
-  if (context.lovell.isLovellVariantPage && context.listing.isListingPage) {
-    return getLovellListingPageStaticPropsResource(
-      resourceType as ListingResourceTypeType,
-      pathInfo,
-      context
-    )
-  }
-
-  // Other Lovell pages depend on base resource
-  const baseResource = (await getDefaultStaticPropsResource(
-    resourceType,
-    pathInfo,
-    context
-  )) as LovellFormattedResource
-
-  // Other: bifurcated pages
-  if (isLovellBifurcatedResource(baseResource)) {
-    return getLovellChildVariantOfResource(
-      baseResource as LovellBifurcatedFormattedResource,
-      context.lovell.variant
-    )
-  }
-
-  // Other: no special treatment
-  return baseResource
 }
 
 export async function getStaticPropsResource(
